Refactor HtmlInstructions to ConfigHtmlInstructions and update related usages in SimpleGalaxyGateConfig and AutobuyConfig
